+++ Semana_3/gobarber-native/gobarberapp/src/pages/Dashboard/index.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import { Text, View, Button } from "react-native";
+
+import { useAuth } from "../../hooks/auth";
+
+let Dashboard: React.FC = () => {
+  let { signOut } = useAuth();
+  return (
+    <View
+      style={{
+        flex: 1,
+        paddingTop: 100,
+        justifyContent: "center",
+        alignItems: "center",
+      }}
+    >
+      <Text>Dashboard</Text>
+      <Button title="Sair" onPress={signOut} />
+    </View>
+  );
+};
+
+export default Dashboard;